backend/requirements.txt
# FastAPI and Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6  # For file uploads
pydantic==2.5.3
pydantic-settings==2.1.0

# PyTorch and Computer Vision
torch==2.1.2
torchvision==0.16.2
Pillow==10.2.0

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1

# Logging and Monitoring (Optional)
python-json-logger==2.0.7

# Development Dependencies (Optional)
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0  # For testing
black==24.1.1  # Code formatting
flake8==7.0.0  # Linting
